+++ components-api/src/main/scala/pl/touk/nussknacker/engine/api/db/DbRef.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.api.db
 
 import com.github.tminglei.slickpg.ExPostgresProfile
-import slick.jdbc.{HsqldbProfile, JdbcBackend, JdbcProfile, PostgresProfile}
+import slick.jdbc.{HsqldbProfile, JdbcBackend, JdbcProfile}
 
 final case class DbRef(db: JdbcBackend.Database, profile: NuJdbcProfile)
 
@@ -18,6 +18,5 @@ trait NuJdbcProfile extends JdbcProfile {
 
 }
 
-class NuPostgresProfile(override val schemaName: String)   extends PostgresProfile with NuJdbcProfile
-class NuExPostgresProfile(override val schemaName: String) extends NuPostgresProfile(schemaName) with ExPostgresProfile
-class NuHsqldbProfile(override val schemaName: String)     extends HsqldbProfile with NuJdbcProfile
+class NuPostgresProfile(override val schemaName: String) extends ExPostgresProfile with NuJdbcProfile
+class NuHsqldbProfile(override val schemaName: String)   extends HsqldbProfile with NuJdbcProfile